User news relation added

+++ app/Http/Controllers/NewsUserController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\NewsUser;
+use App\User;
 use Illuminate\Http\Request;
 
 class NewsUserController extends Controller
@@ -23,20 +24,24 @@ class NewsUserController extends Controller
         $newsuser->imageUrl        = $request->imageUrl;
         $newsuser->newsPublishedAt = $request->newsPublishedAt;
         $newsuser->newsContent     = $request->newsContent;
+        $newsuser->user_id         = auth()->id();
         $newsuser->save();
         return response()->json(['success']);
     }
 
     public function index()
     {
-        $myNews = NewsUser::orderBy('id', 'desc')->paginate(9);
+        $user = User::find(auth()->id());
+        $myNews = $user->usernews()->paginate(9);
         return view('newsuser.index', compact('myNews'));
     }
 
     public function delete(Request $request, $id)
     {
         $deleteNews = NewsUser::findOrFail($id);
-        $deleteNews->delete();
+        if($deleteNews->user_id === auth()->id()) {
+            $deleteNews->delete();
+        }
         return redirect()->back();
     }
 }
